feat: move pygsl logic to a separate module
