+++ spyder/plugins/completion/providers/languageserver/client.py
@@ -5,9 +5,9 @@
 # (see spyder/__init__.py for details)
 
 """
-Spyder Language Server Protocol Client implementation.
+Spyder Language Server Client Wrapper.
 
-It uses pygls JsonRPCClient (_SpyderPyglsClient) with a dedicated asyncio
+It uses pygls JsonRPCClient (SpyderPyglsClient) with a dedicated asyncio
 thread for all LSP communications.  All parameters and responses are typed
 lsprotocol objects; requests and notifications are dispatched directly via
 protocol.send_request_async / protocol.notify.
@@ -23,12 +23,10 @@ import os
 import os.path as osp
 import pathlib
 import sys
-from typing import Optional, TypedDict, TYPE_CHECKING, Union
+from typing import TypedDict, TYPE_CHECKING
 
 # Third-party imports
 from lsprotocol import types as lsp
-from pygls.client import JsonRPCClient
-from pygls.protocol import LanguageServerProtocol, default_converter
 from qtpy.QtCore import QObject, QProcess, Signal, Slot
 from spyder_kernels.utils.pythonenv import is_conda_env
 
@@ -38,11 +36,11 @@ from spyder.api.config.mixins import SpyderConfigurationAccessor
 from spyder.config.base import (
     DEV, get_conf_path, get_debug_level, running_under_pytest
 )
-from spyder.plugins.completion.api import SpyderSymbolKind
 from spyder.plugins.completion.providers.languageserver.decorators import (
     class_register,
     handles,
 )
+from spyder.plugins.completion.providers.languageserver.pygls_client import SpyderPyglsClient
 from spyder.plugins.completion.providers.languageserver.providers import (
     LSPMethodProviderMixIn
 )
@@ -71,240 +69,9 @@ logger = logging.getLogger(__name__)
 _LSP_LOOP = 'lsp'
 
 
-def _spyder_converter():
-    """
-    Build the cattrs converter used by _SpyderPyglsClient.
-
-    Extends pygls's default_converter() with a structure hook for
-    ``Optional[Union[str, NotebookDocumentFilter*]]``.  That type is used by
-    ``NotebookDocumentFilterWithCells.notebook`` in lsprotocol 2025.0.0, but
-    its hooks only cover the non-Optional variant, causing a
-    ``StructureHandlerNotFoundError`` when pylsp returns
-    ``notebookDocumentSync`` capabilities.
-    """
-    converter = default_converter()
-
-    # Build the exact Optional Union type that lsprotocol 2025.0.0 exposes on
-    # NotebookDocumentFilterWithCells.notebook so the hook key matches.
-    _opt_notebook_filter = Optional[
-        Union[
-            str,
-            lsp.NotebookDocumentFilterNotebookType,
-            lsp.NotebookDocumentFilterScheme,
-            lsp.NotebookDocumentFilterPattern,
-        ]
-    ]
-
-    def _structure_opt_notebook_filter(obj, _):
-        if obj is None:
-            return None
-        if isinstance(obj, str):
-            return obj
-        if "notebookType" in obj:
-            return converter.structure(
-                obj, lsp.NotebookDocumentFilterNotebookType
-            )
-        if "scheme" in obj:
-            return converter.structure(obj, lsp.NotebookDocumentFilterScheme)
-        return converter.structure(obj, lsp.NotebookDocumentFilterPattern)
-
-    converter.register_structure_hook(
-        _opt_notebook_filter, _structure_opt_notebook_filter
-    )
-
-    # Register hook for custom for Spyder extended LSP SymbolKind
-    def _structure_symbol_kind(obj, _):
-        try:
-            return lsp.SymbolKind(obj)
-        except ValueError:
-            return SpyderSymbolKind(obj)
-
-    converter.register_structure_hook(lsp.SymbolKind, _structure_symbol_kind)
-
-    return converter
-
-
-def _build_client_capabilities() -> lsp.ClientCapabilities:
-    """Build the ClientCapabilities object advertised to the LSP server."""
-    return lsp.ClientCapabilities(
-        workspace=lsp.WorkspaceClientCapabilities(
-            apply_edit=True,
-            workspace_edit=lsp.WorkspaceEditClientCapabilities(
-                document_changes=True,
-                resource_operations=[
-                    lsp.ResourceOperationKind.Create,
-                    lsp.ResourceOperationKind.Rename,
-                    lsp.ResourceOperationKind.Delete,
-                ],
-                failure_handling=lsp.FailureHandlingKind.Transactional,
-            ),
-            did_change_configuration=lsp.DidChangeConfigurationClientCapabilities(
-                dynamic_registration=True,
-            ),
-            did_change_watched_files=lsp.DidChangeWatchedFilesClientCapabilities(
-                dynamic_registration=True,
-            ),
-            symbol=lsp.WorkspaceSymbolClientCapabilities(
-                dynamic_registration=True,
-            ),
-            execute_command=lsp.ExecuteCommandClientCapabilities(
-                dynamic_registration=True,
-            ),
-            workspace_folders=True,
-            configuration=True,
-        ),
-
-        text_document=lsp.TextDocumentClientCapabilities(
-            synchronization=lsp.TextDocumentSyncClientCapabilities(
-                dynamic_registration=True,
-                will_save=True,
-                will_save_wait_until=True,
-                did_save=True,
-            ),
-            completion=lsp.CompletionClientCapabilities(
-                dynamic_registration=True,
-                completion_item=(
-                    lsp.ClientCompletionItemOptions(
-                        snippet_support=True,
-                        documentation_format=[lsp.MarkupKind.PlainText],
-                    )
-                ),
-            ),
-            hover=lsp.HoverClientCapabilities(
-                dynamic_registration=True,
-                content_format=[lsp.MarkupKind.PlainText],
-            ),
-            signature_help=lsp.SignatureHelpClientCapabilities(
-                dynamic_registration=True,
-                signature_information=(
-                    lsp.ClientSignatureInformationOptions(
-                        documentation_format=[lsp.MarkupKind.PlainText],
-                    )
-                ),
-            ),
-            references=lsp.ReferenceClientCapabilities(
-                dynamic_registration=True,
-            ),
-            document_highlight=lsp.DocumentHighlightClientCapabilities(
-                dynamic_registration=True,
-            ),
-            document_symbol=lsp.DocumentSymbolClientCapabilities(
-                dynamic_registration=True,
-            ),
-            formatting=lsp.DocumentFormattingClientCapabilities(
-                dynamic_registration=True,
-            ),
-            range_formatting=lsp.DocumentRangeFormattingClientCapabilities(
-                dynamic_registration=True,
-            ),
-            on_type_formatting=lsp.DocumentOnTypeFormattingClientCapabilities(
-                dynamic_registration=True,
-            ),
-            definition=lsp.DefinitionClientCapabilities(
-                dynamic_registration=True,
-            ),
-            code_action=lsp.CodeActionClientCapabilities(
-                dynamic_registration=True,
-            ),
-            code_lens=lsp.CodeLensClientCapabilities(
-                dynamic_registration=True,
-            ),
-            document_link=lsp.DocumentLinkClientCapabilities(
-                dynamic_registration=True,
-            ),
-            rename=lsp.RenameClientCapabilities(
-                dynamic_registration=True,
-            ),
-        ),
-    )
-
-
-class _SpyderPyglsClient(JsonRPCClient):
-    """
-    pygls JsonRPCClient subclass that routes server-initiated notifications
-    and requests back to the Qt-thread LSPClient via AsyncDispatcher.
-    """
-
-    def __init__(
-        self, qt_client: "LSPClient", name: str, version: str
-    ) -> None:
-        self.name = name
-        self.version = version
-        super().__init__(
-            protocol_cls=LanguageServerProtocol,
-            converter_factory=_spyder_converter,
-        )
-        self._qt_client = qt_client
-        self._register_server_handlers()
-
-    def _register_server_handlers(self) -> None:
-        """Register handlers for server-to-client messages."""
-        qt = self._qt_client
-
-        # ---- Notifications (server -> client, no response required)
-        @self.feature(lsp.TEXT_DOCUMENT_PUBLISH_DIAGNOSTICS)
-        def on_diagnostics(params: lsp.PublishDiagnosticsParams) -> None:
-            qt._post_notification(
-                lsp.TEXT_DOCUMENT_PUBLISH_DIAGNOSTICS, params
-            )
-
-        @self.feature(lsp.WINDOW_SHOW_MESSAGE)
-        def on_show_message(params: lsp.ShowMessageParams) -> None:
-            qt._post_notification(
-                lsp.WINDOW_SHOW_MESSAGE, params
-            )
-
-        @self.feature(lsp.WINDOW_LOG_MESSAGE)
-        def on_log_message(params: lsp.LogMessageParams) -> None:
-            qt._post_notification(
-                lsp.WINDOW_LOG_MESSAGE, params
-            )
-
-        # ---- Requests from server (require a response)
-        @self.feature(lsp.WORKSPACE_WORKSPACE_FOLDERS)
-        def on_workspace_folders(params: None) -> list:
-            """Return the currently open workspace folders."""
-            folders = []
-            for folder_name, folder_data in qt.watched_folders.items():
-                folders.append(
-                    lsp.WorkspaceFolder(
-                        uri=folder_data['uri'],
-                        name=folder_name,
-                    )
-                )
-            return folders
-
-        @self.feature(lsp.WORKSPACE_CONFIGURATION)
-        def on_workspace_configuration(
-            params: lsp.ConfigurationParams,
-        ) -> list:
-            """Return workspace configuration items requested by the server."""
-            return [qt.configurations] * len(params.items)
-
-        @self.feature(lsp.WORKSPACE_APPLY_EDIT)
-        def on_apply_edit(
-            params: lsp.ApplyWorkspaceEditParams,
-        ) -> lsp.ApplyWorkspaceEditResult:
-            """
-            Route edit application to the Qt thread; acknowledge immediately.
-            """
-            qt._post_notification(
-                lsp.WORKSPACE_APPLY_EDIT, params
-            )
-            return lsp.ApplyWorkspaceEditResult(applied=True)
-
-        @self.feature(lsp.CLIENT_REGISTER_CAPABILITY)
-        def on_register_capability(params: lsp.RegistrationParams) -> None:
-            qt._post_notification(
-                lsp.CLIENT_REGISTER_CAPABILITY, params
-            )
-
-
 @class_register
 class LSPClient(QObject, LSPMethodProviderMixIn, SpyderConfigurationAccessor):
     """
-    Language Server Protocol v3.0 client.
-
     Wraps a pygls BaseLanguageClient running in a dedicated asyncio thread.
     """
 
@@ -369,14 +136,14 @@ class LSPClient(QObject, LSPMethodProviderMixIn, SpyderConfigurationAccessor):
         self._server_cmd = server_settings['cmd']
 
         # LSP capabilities
-        self.client_capabilites = _build_client_capabilities()
+        self.client_capabilites = self._build_client_capabilities()
         self.server_capabilites: lsp.ServerCapabilities | None = None
 
         # QProcess for TCP-mode server management (not used in stdio mode)
         self.server: QProcess | None = None
 
         # pygls client
-        self._pygls_client = _SpyderPyglsClient(
+        self._pygls_client = SpyderPyglsClient(
             self, f"spyder-{self.language}", "0.1.0"
         )
 
@@ -797,4 +564,99 @@ class LSPClient(QObject, LSPMethodProviderMixIn, SpyderConfigurationAccessor):
         self._async_notification(
             lsp.WORKSPACE_DID_CHANGE_CONFIGURATION,
             lsp.DidChangeConfigurationParams(settings=configurations),
+        )
+
+    @staticmethod
+    def _build_client_capabilities() -> lsp.ClientCapabilities:
+        """Build the ClientCapabilities object advertised to the LSP server."""
+        return lsp.ClientCapabilities(
+            workspace=lsp.WorkspaceClientCapabilities(
+                apply_edit=True,
+                workspace_edit=lsp.WorkspaceEditClientCapabilities(
+                    document_changes=True,
+                    resource_operations=[
+                        lsp.ResourceOperationKind.Create,
+                        lsp.ResourceOperationKind.Rename,
+                        lsp.ResourceOperationKind.Delete,
+                    ],
+                    failure_handling=lsp.FailureHandlingKind.Transactional,
+                ),
+                did_change_configuration=lsp.DidChangeConfigurationClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                did_change_watched_files=lsp.DidChangeWatchedFilesClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                symbol=lsp.WorkspaceSymbolClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                execute_command=lsp.ExecuteCommandClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                workspace_folders=True,
+                configuration=True,
+            ),
+
+            text_document=lsp.TextDocumentClientCapabilities(
+                synchronization=lsp.TextDocumentSyncClientCapabilities(
+                    dynamic_registration=True,
+                    will_save=True,
+                    will_save_wait_until=True,
+                    did_save=True,
+                ),
+                completion=lsp.CompletionClientCapabilities(
+                    dynamic_registration=True,
+                    completion_item=(
+                        lsp.ClientCompletionItemOptions(
+                            snippet_support=True,
+                            documentation_format=[lsp.MarkupKind.PlainText],
+                        )
+                    ),
+                ),
+                hover=lsp.HoverClientCapabilities(
+                    dynamic_registration=True,
+                    content_format=[lsp.MarkupKind.PlainText],
+                ),
+                signature_help=lsp.SignatureHelpClientCapabilities(
+                    dynamic_registration=True,
+                    signature_information=(
+                        lsp.ClientSignatureInformationOptions(
+                            documentation_format=[lsp.MarkupKind.PlainText],
+                        )
+                    ),
+                ),
+                references=lsp.ReferenceClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                document_highlight=lsp.DocumentHighlightClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                document_symbol=lsp.DocumentSymbolClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                formatting=lsp.DocumentFormattingClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                range_formatting=lsp.DocumentRangeFormattingClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                on_type_formatting=lsp.DocumentOnTypeFormattingClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                definition=lsp.DefinitionClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                code_action=lsp.CodeActionClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                code_lens=lsp.CodeLensClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                document_link=lsp.DocumentLinkClientCapabilities(
+                    dynamic_registration=True,
+                ),
+                rename=lsp.RenameClientCapabilities(
+                    dynamic_registration=True,
+                ),
+            ),
         )
